Display milestone due date for ticket if due date not set
If a ticket doesn't have a due date but it does have a milestone, show
the milestone's due date. Added `Ticket.due` property that selects the
appropriate due date according to that logic. Note that this only
affects the *display* of due dates. It would be nice if we could also
filter by milestone due dates.