Add milestone due dates to due-on filter for tickets
For tickets that don't have an explicitly set due date but that are
attached to a milestone, include their milestone due dates in the Due
On filter list. If a ticket has both an explicitly set due date and a
milestone due date, only the explicit due date will be included.

This filter now also ignores time completely.